@@ -1,8 +1,26 @@
 module github.com/yoursunny/ndn-prefix-reach
 
-go 1.16
+go 1.18
 
 require (
+	github.com/gorilla/handlers v1.5.1
+	github.com/gorilla/schema v1.2.0
 	github.com/rocketlaunchr/https-go v0.0.0-20200218083740-ba6c48f29f4d
-	github.com/usnistgov/ndn-dpdk v0.0.0-20210518181720-4d388636020f
+	github.com/usnistgov/ndn-dpdk v0.0.0-20220506173138-38934d810545
+	golang.org/x/exp v0.0.0-20220428152302-39d4317da171
+)
+
+require (
+	github.com/felixge/httpsnoop v1.0.1 // indirect
+	github.com/gogf/greuse v1.1.0 // indirect
+	github.com/graphql-go/graphql v0.8.0 // indirect
+	github.com/pascaldekloe/name v1.0.1 // indirect
+	github.com/segmentio/fasthash v1.0.3 // indirect
+	github.com/sethvargo/go-retry v0.2.3 // indirect
+	github.com/tul/emission v0.0.0-20180606124623-7d2aae804ca2 // indirect
+	github.com/zyedidia/generic v1.0.0 // indirect
+	go.uber.org/atomic v1.9.0 // indirect
+	go4.org v0.0.0-20201209231011-d4a079459e60 // indirect
+	golang.org/x/sys v0.0.0-20220503163025-988cb79eb6c6 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 )
